Reprint Server: Canonicalize the Reprint Server WordPress admin surface.
